[refactor] seperate common Layout component directory
