+++ src/components/common/index.ts
@@ -1,7 +1,4 @@
-export * from './Layout';
-export * from './Header';
-export * from './Footer';
-export * from './Menu';
+export * from './layout';
 export * from './selector';
 export * from './toast';
 export * from './modal';
