+++ src/styles/components/common/layout/footer.styles.ts
